unify_auth - merge all the AuthBase classes. common.py both become stubs for (temporary) backwards compatibility. hoist log_response() into lib.py; update client; compat shims in BaseAuth. refactor logger for providers. patch some tests that other changes broke.
